requirements.txt
google-generativeai
requests
transformers
torch
python-dotenv
ratelimit
textblob